Include duration in stop distance calculation
Change bus stop location deduction algorithm to include duration when
deciding how far to render the next stop.
Has a problem when a stop has already been rendered in
one tree branch that it's location has decided to be close to one stop
but really far from another (See distances between A and C and D and C).
Not sure if it is even possible to make it work