Convert index.js into Typescript
This required changing a reference in webpack.common.js to
refer to the generated index file. Removing the file extension seemed
to work. Its also weird that the javascript file worked fine with .js
file extension even though it seemed to actually be a .jsx file.
It did not work as .ts file.