Convert BusStop into Typescript
Some type juggling magic had to be used: get the type from
the variables provided by redux and then combine them to the variables
coming expected to be provided to this class. Feeling good about the
fact that I didn't have to convert this class based component into a
function based even though the instructions were only for the functional
components!