Move input usability check into RouteCalculator
The plan here is to use this check during the validation: if a start
or destination stop is empty, it will not be considered invalid. The
validations still need to be ran entirely so that if we have some input
for one of the stops, it should be still checked for incorrect values.
Like at the moment the validations are ran only when both fields have
input.