Update React and i18n
Had to update i18n too because it ran into some Typescript errors with
translation responses. There was also an issue with multiple values as
children in RouteResult. Fixed this by using a function instead. This
just updates the dependencies. Must follow the React migration guide
to get the full benefits of this upgrade