Start using Redux Tookit
Encountered 2 major issues with the state and broke
RouteResult.test.tsx test. Need to start resolving these next
The first major issue is the calculated route: it cannot be in state as
map.
It is too heavy to serialize like this.
The second is having anchor element of language in the state. It too
needs to be stored somewhere else. It could be a part of local state
even though I've always thought that everything should be centralized in
Redux. I might be wrong