Add test
The test does not work if I don't use stop names from the original
transportData.json. Need to move it to state one day so that I can add
values to it. This is really good experience to encounter these problems
and in future mock them so that testing is easier